requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
Authlib==1.0.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.0.5
django-celery-beat==2.3.0
django-filter==22.1
django-render-block==0.9.1
django-templated-email==3.0.0
django-templated-mail==1.1.1
django-timezone-field==5.0
djangorestframework==3.13.1
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
idna==3.3
importlib-metadata==4.11.4
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
Markdown==3.3.7
MarkupSafe==2.1.1
oauthlib==3.2.0
packaging==21.3
prompt-toolkit==3.0.30
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2022.1
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
zipp==3.8.0
